# openapi.yaml
openapi: "3.0.3"
info:
  title: Secure Multi-Tier Platform API
  description: |
    Enterprise-grade SaaS API backend demonstrating multi-tier AWS architecture.
    Provides product catalogue management, order processing, user authentication
    via AWS Cognito (OAuth2 with PKCE), and platform health monitoring.
  version: "1.0.0"
  contact:
    name: Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development (Docker Compose)
  - url: https://api.example.com
    description: Demo deployment (API Gateway + CloudFront)
...
security:
  - BearerAuth: []
paths:
...
  /orders:
    post:
      operationId: createOrder
      summary: Create a new order
      description: |
        Creates an order after validating stock availability for all items.
        Decrements product stock on success. Requires authenticated user with
        write permission.
      tags:
        - orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
            example:
              items:
                - product_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  quantity: 2
                - product_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                  quantity: 1
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid order (product not found or insufficient stock)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

    get:
      operationId: listOrders
      summary: List user's orders
      description: |
        Returns all orders belonging to the authenticated user, ordered by
        creation date descending.
      tags:
        - orders
      responses:
        "200":
          description: List of user's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{order_id}:
    get:
      operationId: getOrder
      summary: Get order by ID
      description: |
        Returns a specific order if it belongs to the authenticated user.
      tags:
        - orders
      parameters:
        - name: order_id
          in: path
          required: true
          description: Order UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from AWS Cognito via OAuth2 Authorization Code
        flow with PKCE. Include the token in the Authorization header as:
        `Authorization: Bearer <token>`
  schemas:
...
    OrderItemCreate:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
          description: UUID of the product to order
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          description: Quantity to order
    OrderCreate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/OrderItemCreate"
          description: List of items to include in the order
    OrderItemResponse:
      type: object
      required:
        - id
        - product_id
        - quantity
        - unit_price_pence
      properties:
        id:
          type: string
          format: uuid
          description: Order item UUID
        product_id:
          type: string
          format: uuid
          description: Product UUID
        quantity:
          type: integer
          description: Quantity ordered
        unit_price_pence:
          type: integer
          description: Unit price at time of order in pence
    OrderResponse:
      type: object
      required:
        - id
        - user_id
        - status
        - total_pence
        - items
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Order UUID
        user_id:
          type: string
          format: uuid
          description: UUID of the user who placed the order
        status:
          type: string
          description: Order status
          example: pending
        total_pence:
          type: integer
          description: Total order value in pence
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
          description: Line items in the order
        created_at:
          type: string
          format: date-time
          description: Order creation timestamp

    # --- Errors ---
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "INVALID_CREDENTIALS"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid email or password"
            request_id:
              type: string
              nullable: true
              description: Request correlation ID for tracing
            details:
              type: object
              nullable: true
              additionalProperties: true
              description: Additional error context
    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Field location path
              msg:
                type: string
                description: Validation error message
              type:
                type: string
                description: Error type identifier
          description: List of field-level validation errors